pyretic_firewall.py created online with Bitbucket
In this assignment, the task is to implement a layer 2 firewall that runs alongside the MAC learning module on the Pyretic runtime. 
The firewall application is provided with a list of MAC address pairs i.e., access control list (ACLs). 
When a connection establishes between the controller and the switch, the application installs static flow rule entries in the OpenFlow table to disable all communication between each MAC pair.