NetASM: A Network Assembly for Orchestrating Programmable Network Devices
NetASM is analogous to an x86 or MIPS-like instructions set.  However, unlike updating main memory and registers, it defines the layout (i.e., topology and states) of the data plane. Topology defines how the packet is traversed through the data plane, and state refers to the type of memory element (i.e., register or table). These state elements have a well-defined data structure and type declaration in NetASM, which makes it easy to identify bugs early in the compilation process.

Stateless firewall assignement task:
- Create a new ACL table using the MKT instruction and update the initialization code.
- Update the topology code using the topology instructions BRTF and DRP, such that only communication between the MAC pair entries listed in the ACL table are allowed.
- Write firewall rules in the ACL table with allowed MAC pairs, using the control instruction WRT.